Crate monocypher

Monocypher is a cryptographic library.

It provides functions for authenticated encryption, hashing, password key derivation, key exchange, and public key signatures.

Visit the official documentation for details.

Modules

argon2i
Argon2i key derivation function

blake2b
Blake2b hash function

chacha20
Chacha20 encryption functions.

check
Public key signature functions

key_exchange
Elliptic Curve Diffie-Hellman key exchange

lock
Authenticated encryption w/o additional data

poly1305
Poly1305 one-time message authentication codes.

sign
Incremental public key signatures

unlock
Authenticated decryption w/o additional data

utils
Util functions for constant time comparison and memory wiping.